requirements.txt
firebase-admin
google-generativeai
tqdm
